+++ R/conditions.R
@@ -1659,6 +1659,45 @@ intellectualDisability_list <- function(dMeasure_obj, appointments = NULL) {
     unique()
 })
 
+#' list of patients with history of refugee or asylum seeker
+#'
+#' @param dMeasure_obj dMeasure R6 object
+#' @param appointments dataframe of appointments $InternalID and $Date
+#'
+#'  if no parameter provided, derives from $appointments_filtered
+#'
+#' @return a vector of numbers, which are the InternalIDs
+#' @export
+refugeeAsylum_list <- function(dMeasure_obj, appointments = NULL) {
+  dMeasure_obj$refugeeAsylum_list(appointments)
+}
+.public(dMeasure, "refugeeAsylum_list", function(appointments = NULL) {
+  # @param Appointments dataframe of $InternalID and $Date
+  #  if no parameter provided, derives from $appointments_filtered
+  #
+  # Returns vector of InternalID of patients who have intellectual disability
+
+  if (is.null(appointments)) {
+    appointments <- self$appointments_filtered %>>%
+      dplyr::select(InternalID, AppointmentDate) %>>%
+      dplyr::rename(Date = AppointmentDate)
+    # just needs $InternalID and $Date
+  }
+
+  intID <- c(dplyr::pull(appointments, InternalID), -1)
+  # internalID in appointments. add a -1 in case this is an empty list
+
+  self$db$history %>>%
+    dplyr::filter(
+      ConditionID %in% c(13154, 13155),
+      # 13154 = refugee
+      # 13155 = asylum seeker
+      InternalID %in% intID
+    ) %>>%
+    dplyr::pull(InternalID) %>>%
+    unique()
+})
+
 #' list of patients who are listed as the head of family
 #' with a child between defined ages
 #'